+++ deafrica_conflux/__version__.py
@@ -1,4 +1,4 @@
 # file generated by setuptools_scm
 # don't change, don't track in version control
-__version__ = version = '0.1.dev401+g911e5a3.d20230803'
-__version_tuple__ = version_tuple = (0, 1, 'dev401', 'g911e5a3.d20230803')
+__version__ = version = '0.1.dev433+gbf52fb0.d20230824'
+__version_tuple__ = version_tuple = (0, 1, 'dev433', 'gbf52fb0.d20230824')
