+++ deafrica_conflux/plugins/__init__.py
@@ -1,3 +1,3 @@
-from ._plugin_utils import run_plugin, validate_plugin
+from .utils import run_plugin, validate_plugin
 
 __all__ = ["run_plugin", "validate_plugin"]
